server/requirements.txt
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
coloredlogs==15.0.1
cryptography==44.0.0
firebase-admin==6.6.0
Flask==3.1.0
flatbuffers==24.3.25
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
numpy==2.2.0
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
requests==2.32.3
rsa==4.9
sympy==1.13.3
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
